Module Tezos_benchmark_type_inference_alpha.Mikhailsky

Mikhailsky: Michelson in Micheline form, with typed holes and annotations. Mikhailsky terms are hash-consed.

Michelson code is a hard to type-check and generate incrementally due to the presence of ambiguous constructs, such as literals like { 1 ; 2 ; 3 }. Is it a list of ints? of nats? of tez? Or a set?

Thus, we will work with Mikhailsky, a better behaved version of Michelson allowing local reconstruction of types.

Differences wrt Michelson:

1. non string/byte literals are explicitly annotated with their head type constructor. Here is an int i: Prim (_, D_int, Int i, _) Here is an nat n: Prim (_, D_nat, Int i, _) Here is an list of something: Prim (_, D_list, michelson_list, _) Here is a set: Prim (_, D_set, michelson_set, _) Here is a map: Prim (_, D_map, michelson_map, _) etc. Projecting back from this language to Michelson is trivial.

2. Instructions `LEFT/RIGHT` do not need to carry the type of the other component of the disjunction. These has to be filled in back when generating Michelson from Mikhailsky.

4. The same holds for the input/output type of a lambda as specified in the `LAMBDA` instruction.

3. Some instructions are annotated with the type on which they operate. Eg if Prim (_, I_ADD, , ) is the (ad-hoc polymorphic) addition in Michelson, we will have the following variants in Mikhailsky:

Prim (_, I_ADD, Prim (_, T_mutez, [], []), Prim (_, T_mutez, [], []) , ) for mutez addition

Prim (_, I_ADD, Prim (_, T_int, [], []), Prim (_, T_nat, [], []) , ) for int+nat addition etc.

module Mikhailsky_signature : Tezos_micheline_rewriting.Algebraic_signature.S with type t = Mikhailsky_prim.prim

The signature of Mikhailsky terms.

module Path : Tezos_micheline_rewriting.Path.S

Elements of type Path.t allow to index subterms of Mikhailsky terms.

type label = Tezos_micheline_rewriting.Micheline_with_hash_consing.hcons_info

The following types correspond to those provided when instantiating the functor Micheline_with_hash_consing.Make on Mikhailsky_signature.

type head = Mikhailsky_signature.t

type node = (label, head) Tezos_micheline.Micheline.node

exception Term_contains_holes

exception Ill_formed_mikhailsky

val parse_ty : allow_big_map:bool -> allow_operation:bool -> allow_contract:bool -> node -> Type.Base.t

parse_ty returns a type from a Mikhailsky term.

val map_var : (int -> node) -> Type.Base.t -> node

map_var f x maps the function f on all variables contained in the type x.

val unparse_ty_exn : Type.Base.t -> node

unparse_ty returns a Mikhailsky term representing a type.

val unparse_ty : Type.Base.t -> node option

val to_michelson : node -> Tezos_protocol_alpha.Protocol.Script_repr.expr

Extracts a Michelson term from a Mikhailsky one. Raises Term_contains_holes if it cannot be done.

val pp : Stdlib.Format.formatter -> node -> unit

Pretty printer.

val to_string : node -> string

val size : node -> int

Returns the number of nodes of a Mikhailsky term.

val prim : Mikhailsky_prim.prim -> node list -> string list -> node

Micheline generic constructors

val seq : node list -> node

val string : string -> node

val bytes : Stdlib.Bytes.t -> node

Mikhailsky smart constructors

val instr_hole : node

Holes

val data_hole : node

val unit_ty : node

Types

val int_ty : node

val nat_ty : node

val bool_ty : node

val string_ty : node

val bytes_ty : node

val key_hash_ty : node

val option_ty : node -> node

val list_ty : node -> node

val tag : node -> int

Project unique tag out of Mikhailsky node

val hash : node -> int

Project hash out of Mikhailsky node

module Instructions : sig ... end

Instructions

module Data : sig ... end

data